Added more reusable items to the project, enabling the gear page to be more streamlined
