+++ bike-banter/src/app/components/molecules/bike-components.tsx
@@ -1,4 +1,3 @@
-
 export default function BikeComponents() {
     return(
         <div className="bg-muted/50 aspect-video rounded-xl">
